+++ mesh/CMakeLists.txt
@@ -1,0 +1,21 @@
+# SPDX-License-Identifier: Apache-2.0
+
+cmake_minimum_required(VERSION 3.20.0)
+set(QEMU_EXTRA_FLAGS -s)
+
+set(BOARD thingy52_nrf52832)
+
+if((BOARD STREQUAL nrf51_blenano) OR (BOARD STREQUAL nrf51_ble400))
+  set(CONF_FILE nrf51_qfaa.conf)
+endif()
+
+find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
+project(mesh)
+
+target_sources(app PRIVATE src/main.c)
+
+if (BOARD STREQUAL bbc_microbit)
+  target_sources(app PRIVATE src/microbit.c)
+else()
+  target_sources(app PRIVATE src/board.c)
+endif()
