+++ mesh/nrf51_qfaa.conf
@@ -1,0 +1,48 @@
+CONFIG_INIT_STACKS=y
+CONFIG_ISR_STACK_SIZE=512
+CONFIG_MAIN_STACK_SIZE=512
+CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=1024
+
+CONFIG_BT=y
+CONFIG_BT_TINYCRYPT_ECC=y
+CONFIG_BT_RX_STACK_SIZE=1100
+
+CONFIG_BT_OBSERVER=y
+CONFIG_BT_BROADCASTER=y
+CONFIG_BT_CTLR_DUP_FILTER_LEN=0
+
+CONFIG_BT_CTLR_LE_ENC=n
+CONFIG_BT_CTLR_LE_PING=n
+CONFIG_BT_DATA_LEN_UPDATE=n
+CONFIG_BT_PHY_UPDATE=n
+CONFIG_BT_CTLR_CHAN_SEL_2=n
+CONFIG_BT_CTLR_MIN_USED_CHAN=n
+CONFIG_BT_CTLR_ADV_EXT=n
+CONFIG_BT_CTLR_PRIVACY=n
+
+CONFIG_BT_MESH=y
+CONFIG_BT_MESH_LOW_POWER=y
+CONFIG_BT_MESH_ADV_BUF_COUNT=3
+CONFIG_BT_MESH_SEG_BUFS=6
+CONFIG_BT_MESH_RELAY=y
+
+CONFIG_BT_MESH_SUBNET_COUNT=1
+CONFIG_BT_MESH_APP_KEY_COUNT=1
+CONFIG_BT_MESH_MODEL_GROUP_COUNT=1
+CONFIG_BT_MESH_LABEL_COUNT=0
+
+#CONFIG_UART_CONSOLE=n
+#CONFIG_BT_DEBUG_MONITOR_UART=y
+#CONFIG_BT_DEBUG_LOG=y
+
+#CONFIG_BT_MESH_SELF_TEST=y
+#CONFIG_BT_MESH_IV_UPDATE_TEST=y
+
+#CONFIG_BT_MESH_DEBUG=y
+#CONFIG_BT_MESH_DEBUG_NET=y
+#CONFIG_BT_MESH_DEBUG_TRANS=y
+#CONFIG_BT_MESH_DEBUG_PROV=y
+#CONFIG_BT_MESH_DEBUG_BEACON=y
+#CONFIG_BT_MESH_DEBUG_CRYPTO=y
+#CONFIG_BT_MESH_DEBUG_ADV=y
+#CONFIG_BT_MESH_DEBUG_ACCESS=y
