+++ mesh/prj.conf
@@ -28,7 +28,7 @@ CONFIG_BT_MESH_LOW_POWER=n
 CONFIG_BT_MESH_FRIEND=n
 
 CONFIG_BT_MESH_PB_GATT=y
-CONFIG_BT_MESH_PB_ADV=n
+CONFIG_BT_MESH_PB_ADV=y
 CONFIG_BT_MESH_GATT_PROXY=y
 
 CONFIG_BT_MESH_LPN_SCAN_LATENCY=30
@@ -44,12 +44,9 @@ CONFIG_BT_MESH_IV_UPDATE_TEST=y
 
 #CONFIG_UART_CONSOLE=n
 #CONFIG_BT_DEBUG_MONITOR_UART=y
+CONFIG_BT_DEBUG_LOG=y
 
-
-#### set this to y for detailed debug information 
-CONFIG_BT_DEBUG_LOG=n
-
-CONFIG_BT_MESH_DEBUG=y
+CONFIG_BT_MESH_DEBUG=n
 CONFIG_BT_MESH_DEBUG_PROV=y
 CONFIG_BT_MESH_DEBUG_PROXY=y
 #CONFIG_BT_MESH_DEBUG_BEACON=y
@@ -64,9 +61,6 @@ CONFIG_BT_MESH_DEBUG_ADV=y
 #CONFIG_BT_MESH_SELF_TEST=y
 
 
-CONFIG_BT_MESH_CFG_CLI=y
-
-
 
 #-----------------------------RTT_CONFIG--------------------------------------
 CONFIG_USE_SEGGER_RTT=y
@@ -74,3 +68,7 @@ CONFIG_USE_SEGGER_RTT=y
 CONFIG_RTT_CONSOLE=y
 CONFIG_UART_CONSOLE=n
 #-----------------------------------------------------------------------------
+
+
+
+CONFIG_BT_MESH_CFG_CLI=y
