+++ mesh/prj_bbc_microbit.conf
@@ -1,0 +1,63 @@
+CONFIG_INIT_STACKS=y
+CONFIG_ISR_STACK_SIZE=768
+CONFIG_MAIN_STACK_SIZE=512
+CONFIG_DISPLAY=y
+CONFIG_MICROBIT_DISPLAY=y
+CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=1024
+CONFIG_MICROBIT_DISPLAY_STR_MAX=5
+CONFIG_GPIO=y
+
+CONFIG_PM=n
+
+CONFIG_BT_SETTINGS=y
+CONFIG_FLASH=y
+CONFIG_FLASH_PAGE_LAYOUT=y
+CONFIG_FLASH_MAP=y
+CONFIG_NVS=y
+CONFIG_SETTINGS=y
+
+CONFIG_BT=y
+CONFIG_BT_TINYCRYPT_ECC=y
+CONFIG_BT_RX_STACK_SIZE=1100
+CONFIG_BT_BUF_EVT_RX_COUNT=3
+CONFIG_BT_BUF_ACL_RX_COUNT=3
+CONFIG_BT_BUF_EVT_DISCARDABLE_COUNT=3
+
+CONFIG_BT_CTLR_DUP_FILTER_LEN=0
+CONFIG_BT_OBSERVER=y
+CONFIG_BT_BROADCASTER=y
+
+CONFIG_BT_CTLR_LE_ENC=n
+CONFIG_BT_CTLR_LE_PING=n
+CONFIG_BT_DATA_LEN_UPDATE=n
+CONFIG_BT_PHY_UPDATE=n
+CONFIG_BT_CTLR_CHAN_SEL_2=n
+CONFIG_BT_CTLR_MIN_USED_CHAN=n
+CONFIG_BT_CTLR_ADV_EXT=n
+CONFIG_BT_CTLR_PRIVACY=n
+
+CONFIG_BT_MESH=y
+CONFIG_BT_MESH_ADV_BUF_COUNT=3
+CONFIG_BT_MESH_SEG_BUFS=3
+CONFIG_BT_MESH_RELAY=y
+CONFIG_BT_MESH_LOOPBACK_BUFS=1
+#CONFIG_BT_MESH_LOW_POWER=y
+
+#CONFIG_BT_MESH_SELF_TEST=y
+#CONFIG_BT_MESH_IV_UPDATE_TEST=y
+
+CONFIG_BT_MESH_SUBNET_COUNT=1
+CONFIG_BT_MESH_APP_KEY_COUNT=1
+CONFIG_BT_MESH_MODEL_GROUP_COUNT=1
+CONFIG_BT_MESH_LABEL_COUNT=0
+
+#CONFIG_BT_DEBUG_LOG=y
+
+#CONFIG_BT_MESH_DEBUG=y
+#CONFIG_BT_MESH_DEBUG_NET=y
+#CONFIG_BT_MESH_DEBUG_TRANS=y
+#CONFIG_BT_MESH_DEBUG_PROV=y
+#CONFIG_BT_MESH_DEBUG_BEACON=y
+#CONFIG_BT_MESH_DEBUG_CRYPTO=y
+#CONFIG_BT_MESH_DEBUG_ADV=y
+#CONFIG_BT_MESH_DEBUG_ACCESS=y
