--- mesh_sample/CMakeLists.txt
@@ -1,16 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-cmake_minimum_required(VERSION 3.20.0)
-
-
-set(BOARD thingy52_nrf52832)
-
-
-find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(mesh)
-
-target_sources(app PRIVATE src/main.c)
-
-
-target_sources(app PRIVATE src/board.c)
-
