--- mesh_sample/src/ble_mesh.c
@@ -1,4 +1,0 @@
-
-#include "ble_mesh.h"
-#include "device_composition.h"
-
