--- mesh_sample/src/device_composition.c
@@ -1,3 +1,0 @@
-#include <drivers/gpio.h>
-
-#include "device_composition.h"
