+++ node_base/CMakeLists.txt
@@ -1,11 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
-
-set(BOARD thingy52_nrf52832)
-
+set(BOARD nrf52840dongle_nrf52840)
 cmake_minimum_required(VERSION 3.20.0)
-set(QEMU_EXTRA_FLAGS -s)
-
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(mesh_provisioner)
+project(blinky)
 
 target_sources(app PRIVATE src/main.c)
