+++ node_base/prj.conf
@@ -1,3 +1,35 @@
+CONFIG_GPIO=y
+#Enable this for Arduino Nano board
+#CONFIG_BOARD_ARDUINO_NANO_33_BLE_EN_USB_CONSOLE=y
+
+#-----------------------------------------------------------------------------
+CONFIG_USB_DEVICE_STACK=y
+CONFIG_USB_DEVICE_PRODUCT="Zephyr USB console sample"
+
+CONFIG_SERIAL=y
+CONFIG_CONSOLE=y
+CONFIG_UART_CONSOLE=y
+CONFIG_UART_LINE_CTRL=y
+#-----------------------------------------------------------------------------
+
+#--------------------------------USB OPTIONS----------------------------------
+CONFIG_USB_DEVICE_PRODUCT="Particle Argon - Zephyr"
+CONFIG_USB_DEVICE_MANUFACTURER="Wilfred MK"
+CONFIG_USB_DEVICE_VID=0xC553
+CONFIG_USB_DEVICE_PID=0x4011
+#-----------------------------------------------------------------------------
+
+CONFIG_BT=y
+CONFIG_BT_BROADCASTER=y
+CONFIG_BT_OBSERVER=y
+#CONFIG_BT_DEBUG_LOG=y
+CONFIG_BT_DEVICE_NAME="base"
+
+
+
+
+
+
 #CONFIG_INIT_STACKS=y
 CONFIG_MAIN_STACK_SIZE=1408
 CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
@@ -52,12 +84,3 @@ CONFIG_BT_MESH_RPL_STORE_TIMEOUT=600
 #CONFIG_BT_MESH_DEBUG_BEACON=y
 #CONFIG_BT_MESH_DEBUG_CRYPTO=y
 #CONFIG_BT_MESH_DEBUG_ADV=y
-
-
-
-#-----------------------------RTT_CONFIG--------------------------------------
-CONFIG_USE_SEGGER_RTT=y
-
-CONFIG_RTT_CONSOLE=y
-CONFIG_UART_CONSOLE=n
-#-----------------------------------------------------------------------------
